fix: accessToken hiding

+++ src/app/module/utils/moduleCookie.ts
@@ -4,7 +4,7 @@ import { jwtDecode } from 'jwt-decode'
 import { modulePostFetch } from './moduleFetch'
 
 import { API_SUCCESS_CODE, KEY_ACCESS_TOKEN } from '@/constant/constant'
-import { ERR_COOKIE_NOT_FOUND } from '@/constant/errorMsg'
+import { ERR_ACCESS, ERR_COOKIE_NOT_FOUND } from '@/constant/errorMsg'
 import { API_URL_REFRESH } from '@/constant/route/api-route-constant'
 import {
   type ApiResponseType,
@@ -14,12 +14,15 @@ import {
 } from '@/types/module'
 
 export const moduleGetCookie = (name: string) => {
+  if (name === KEY_ACCESS_TOKEN) return ERR_ACCESS
   if (hasCookie(name)) {
     return getCookie(name) as string
   }
+
   return ERR_COOKIE_NOT_FOUND
 }
 export const moduleSetCookies = (cookies: Record<string, string | number>): void => {
+  if (Object.keys(cookies)[0] === KEY_ACCESS_TOKEN) return
   Object.entries(cookies).forEach(([key, value]) => {
     setCookie(key, value)
   })
@@ -27,7 +30,7 @@ export const moduleSetCookies = (cookies: Record<string, string | number>): void
 
 export const moduleDeleteCookies = (...name: string[]): void => {
   name.forEach((name) => {
-    deleteCookie(name)
+    if (name !== KEY_ACCESS_TOKEN) deleteCookie(name)
   })
 }
